Extract ScheduleAddCell component into separate file
